file was becoming too large, so move out type lookups into their own file...
